requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
aiomysql==0.2.0
sqlalchemy 
pymysql
python-dotenv
cryptography